Send ipv4 OK

+++ srcs/display/debug.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/25 16:53:03 by arsciand          #+#    #+#             */
-/*   Updated: 2021/05/25 16:53:57 by arsciand         ###   ########.fr       */
+/*   Updated: 2021/09/07 13:44:01 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void    print_bytes(int bytes, void *msg)
 {
-    for (size_t i = 0; i < bytes; i++)
+    for (int i = 0; i < bytes; i++)
     {
         if (!(i & 15))
             printf("\n%04X:  ", i);
